% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{calculate_cluster_variance_optimized}
\alias{calculate_cluster_variance_optimized}
\title{Optimized cluster variance calculation}
\usage{
calculate_cluster_variance_optimized(cov_matrix, asset_names)

calculate_cluster_variance_optimized(cov_matrix, asset_names)
}
\description{
Optimized cluster variance calculation

Optimized cluster variance calculation
}
\keyword{internal}
