% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{calculate_hrp_weights}
\alias{calculate_hrp_weights}
\title{Calculate HRP weights for a given returns matrix}
\usage{
calculate_hrp_weights(
  returns_matrix,
  asset_names,
  cluster_method,
  distance_method
)
}
\description{
Calculate HRP weights for a given returns matrix
}
\keyword{internal}
