% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{calculate_max_div_weights}
\alias{calculate_max_div_weights}
\title{Calculate Maximum Diversification Portfolio weights}
\usage{
calculate_max_div_weights(cov_matrix, asset_names)
}
\description{
Calculate Maximum Diversification Portfolio weights
}
\keyword{internal}
