% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{demo_sector_map}
\alias{demo_sector_map}
\title{Demo sector map for examples/tests}
\usage{
demo_sector_map(symbols, n_groups = 2L)
}
\arguments{
\item{symbols}{character vector.}

\item{n_groups}{integer groups to assign.}
}
\description{
Demo sector map for examples/tests
}
