% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walk_forward.R
\name{.wf_make_splits}
\alias{.wf_make_splits}
\title{Create Window Splits for Walk-Forward}
\usage{
.wf_make_splits(n, is_periods, oos_periods, step)
}
\description{
Create Window Splits for Walk-Forward
}
\keyword{internal}
