% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensure_dt_copy}
\alias{ensure_dt_copy}
\title{Ensure Data.Table Without Mutation}
\usage{
ensure_dt_copy(data)
}
\arguments{
\item{data}{Data.frame or data.table}
}
\value{
Copy of data as data.table
}
\description{
Converts input to data.table if needed, always returning a copy
to prevent accidental data mutation. Core safety function used
throughout the library.
}
\examples{
data("sample_prices_weekly")
dt <- ensure_dt_copy(sample_prices_weekly)  # Safe to modify dt
}
