% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_above}
\alias{filter_above}
\title{Filter Stocks Above Threshold}
\usage{
filter_above(signal_df, value)
}
\arguments{
\item{signal_df}{Data frame with signal values}

\item{value}{Threshold value}
}
\value{
Binary selection matrix
}
\description{
Convenience function to select stocks with signal above a value.
}
\examples{
# Select stocks with RSI > 70
overbought <- filter_above(rsi, 70)
}
