% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_below}
\alias{filter_below}
\title{Filter Stocks Below Threshold}
\usage{
filter_below(signal_df, value)
}
\arguments{
\item{signal_df}{Data frame with signal values}

\item{value}{Threshold value}
}
\value{
Binary selection matrix
}
\description{
Convenience function to select stocks with signal below a value.
}
\examples{
data("sample_prices_weekly")
rsi <- calc_rsi(sample_prices_weekly, 14)
oversold <- filter_below(rsi, 30)
}
