% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_analytics.R
\name{get_data_frequency}
\alias{get_data_frequency}
\title{Detect Data Frequency from Dates}
\usage{
get_data_frequency(dates)
}
\arguments{
\item{dates}{Vector of Date objects}
}
\value{
Character string: "daily", "weekly", "monthly", or "quarterly"
}
\description{
Automatically detects whether data is daily, weekly, monthly, or
quarterly based on date spacing.
}
\examples{
data("sample_prices_weekly")
freq <- get_data_frequency(sample_prices_weekly$Date)
}
