% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{join_panels}
\alias{join_panels}
\title{Join multiple panels on intersecting dates (unique symbol names)}
\usage{
join_panels(panels)
}
\arguments{
\item{panels}{list of wide panels \code{Date + symbols}.}
}
\description{
Join multiple panels on intersecting dates (unique symbol names)
}
