% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_adapters.R
\name{manual_adapter}
\alias{manual_adapter}
\title{Adapter for User-Provided Data}
\usage{
manual_adapter(data, date_col = "Date")
}
\arguments{
\item{data}{User-provided data frame}

\item{date_col}{Name of date column (default: "Date")}
}
\value{
Standardized data.table
}
\description{
Simple adapter for when users provide their own data frame.
Ensures proper Date formatting and sorting.
}
\examples{
# Use your own data frame
data("sample_prices_weekly")
my_prices <- manual_adapter(sample_prices_weekly)
}
