% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{panel_lag}
\alias{panel_lag}
\title{Lag all symbol columns by k}
\usage{
panel_lag(df, k = 1L)
}
\arguments{
\item{df}{wide panel with \code{Date + symbols}.}

\item{k}{integer lag.}
}
\description{
Lag all symbol columns by k
}
