% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortfolioTesteR-package.R
\docType{package}
\name{PortfolioTesteR-package}
\alias{PortfolioTesteR}
\alias{PortfolioTesteR-package}
\title{PortfolioTesteR: Test Investment Strategies with English-Like Code}
\description{
Tools to design, backtest, and analyze portfolio strategies using simple, English-like function chains. Includes momentum, RSI, regime filters, position limits, and multiple weighting schemes with plotting and summaries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alb3rtazzo/PortfolioTesteR}
  \item Report bugs at \url{https://github.com/alb3rtazzo/PortfolioTesteR/issues}
}

}
\keyword{internal}
