% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{print.param_grid_result}
\alias{print.param_grid_result}
\title{Print a param_grid_result}
\usage{
\method{print}{param_grid_result}(x, ...)
}
\arguments{
\item{x}{A \code{param_grid_result} object returned by \code{\link[=run_param_grid]{run_param_grid()}}.}

\item{...}{Additional arguments passed to methods (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a param_grid_result
}
