% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walk_forward.R
\name{print.wf_optimization_result}
\alias{print.wf_optimization_result}
\title{Print a wf_optimization_result}
\usage{
\method{print}{wf_optimization_result}(x, ...)
}
\arguments{
\item{x}{A \code{wf_optimization_result} object returned by \code{\link[=run_walk_forward]{run_walk_forward()}}.}

\item{...}{Additional arguments passed to methods (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a wf_optimization_result
}
