% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{recursive_bisection_optimized}
\alias{recursive_bisection_optimized}
\title{Optimized recursive bisection for HRP}
\usage{
recursive_bisection_optimized(cov_matrix, asset_order)

recursive_bisection_optimized(cov_matrix, asset_order)
}
\description{
Optimized recursive bisection for HRP

Optimized recursive bisection for HRP
}
\keyword{internal}
