% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{roll_ic_stats}
\alias{roll_ic_stats}
\title{Rolling IC mean / sd / ICIR}
\usage{
roll_ic_stats(ic_dt, window = 26L)
}
\arguments{
\item{ic_dt}{output of \code{\link[=ic_series]{ic_series()}}.}

\item{window}{rolling window length.}
}
\description{
Rolling IC mean / sd / ICIR
}
