% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtesting.R
\name{summary.backtest_result}
\alias{summary.backtest_result}
\title{Summary method for backtest results}
\usage{
\method{summary}{backtest_result}(object, ...)
}
\arguments{
\item{object}{A backtest_result object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisible copy of the object
}
\description{
Summary method for backtest results
}
