% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{turnover_by_date}
\alias{turnover_by_date}
\title{Turnover by date}
\usage{
turnover_by_date(weights)
}
\arguments{
\item{weights}{weight panel.}
}
\value{
data.table with \verb{Date, turnover} (0.5 * L1 change).
}
\description{
Turnover by date
}
