% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utilities.R
\name{update_vix_in_db}
\alias{update_vix_in_db}
\title{Update VIX data in database}
\usage{
update_vix_in_db(db_path, from_date = NULL)
}
\arguments{
\item{db_path}{Path to SQLite database}

\item{from_date}{Start date for update (NULL = auto-detect)}
}
\value{
Number of rows updated (invisible)
}
\description{
Update VIX data in database
}
