% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{validate_group_map}
\alias{validate_group_map}
\title{Validate or synthesize a group map}
\usage{
validate_group_map(symbols, group_map)
}
\arguments{
\item{symbols}{character vector of symbols.}

\item{group_map}{\code{data.frame(Symbol, Group)}.}
}
\description{
Validate or synthesize a group map
}
