% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PortfolioTesteR-package.R
\docType{package}
\name{PortfolioTesteR-package}
\alias{PortfolioTesteR}
\alias{PortfolioTesteR-package}
\title{PortfolioTesteR: Test Investment Strategies with English-Like Code}
\description{
Design, backtest, and analyze portfolio strategies using simple, English-like function chains. Includes technical indicators, flexible stock selection, portfolio construction methods (equal weighting, signal weighting, inverse volatility, hierarchical risk parity), and a compact backtesting engine for portfolio returns, drawdowns, and summary metrics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AlbertoPallotta/PortfolioTesteR}
  \item Report bugs at \url{https://github.com/AlbertoPallotta/PortfolioTesteR/issues}
}

}
\author{
\strong{Maintainer}: Alberto Pallotta \email{pallottaalberto@gmail.com}

}
\keyword{internal}
