% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtesting.R
\name{print.backtest_result}
\alias{print.backtest_result}
\title{Print Backtest Results}
\usage{
\method{print}{backtest_result}(x, ...)
}
\arguments{
\item{x}{backtest_result object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisible copy of x
}
\description{
S3 print method for backtest results. Shows key performance metrics.
}
\examples{
data("sample_prices_weekly")
mom <- calc_momentum(sample_prices_weekly, lookback = 12)
sel <- filter_top_n(mom, n = 10)
W   <- weight_equally(sel)
res <- run_backtest(sample_prices_weekly, W)
print(res)
}
